pyproject.toml
[tool.poetry]
name = "shareyourcloning-submission"
version = "0.1.0"
description = ""
authors = ["Manuel Lera-Ramirez <manulera14@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
pygithub = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
